Fix failing integration test
The test has failed since the implementation of testsuite properties
overriding when using the test goal.
In this test we consider that the runner type parameter is used by the
mojo when we found logs for SoapuiTestCaseRunner.

With the new feature, we add a new custom TestCaseRunner, so the logs
don't exist anymore.
The fix is to make this logs occur and ajust expected logs for the new
runner.