Test checks that no property is created
In these tests, we check that a property is only updated for the specific
test suite. We already verified that for other suites that already have a
property with the same name, this property is not updated.
Now we also verify that for suites that previously does not have the
property, the given property is not created.