Improve the way we retrieve artifacts #4
soapui jar is now copied in the bin directory instead of being retrieved
from the maven repository. We currently have an issue on Ubuntu with
openjdk 6 and 7, the soapui version cannot be retrieved from the manifest,
so we could not use it to get the soapui jar

Also use our plugin to get all jars. SmartBear plugins may have missing
dependencies (at least prior to the 4.5.2 version), so use our plugin to
fix this