Fix soapui oss mock-as-war failing test #4
The test failed when the repository did not have the maven soapui plugin.
In this case, the plugin is retrived from the soapui repository, but then
maven try to resolve it from other repo. They did not have it and the
build fail!

I did not take time to investage why we have such a behaviour. Installing
the plugin before running the test is a workaround, so this is ok for now.